+++ app/src/main/java/com/eddy/nrf/utils/Uuid.kt
@@ -1,0 +1,16 @@
+package com.eddy.nrf.utils
+
+
+import java.util.UUID
+
+object Uuid {
+
+    /*service*/
+    val HEART_RATE_SERVICE: UUID = UUID.fromString("0000180D-0000-1000-8000-00805F9B34FB")
+
+    /*Characteristic*/
+    var HEART_RATE_MEASUREMENT: UUID = UUID.fromString("00002a37-0000-1000-8000-00805f9b34fb")
+
+    /* Mandatory Client Characteristic Config Descriptor */
+    val CLIENT_CONFIG: UUID = UUID.fromString("00002902-0000-1000-8000-00805f9b34fb")
+}
